src/main/python/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==4.6.0

# Async Tasks
celery[redis]==5.3.4

# Binance API
python-binance==1.0.19
ccxt==4.1.99

# Data Processing & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0
# ta-lib==0.4.28  # Removed: requires system dependencies
yfinance==0.2.33  # For historical data
pandas-ta==0.3.14b0  # Alternative technical analysis library

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
mplfinance==0.12.10b0
Pillow==10.1.0
selenium==4.15.2

# API & Validation
httpx==0.25.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.38.0

# Date & Time
pendulum==2.1.2
pytz==2023.3

# Other Utilities
requests==2.31.0
aiofiles==23.2.1
orjson==3.9.10